+++ packages/gitbook/src/components/RootLayout/CustomizationRootLayout.tsx
@@ -1,8 +1,8 @@
 import {
+    CustomizationDefaultThemeMode,
     CustomizationIconsStyle,
     CustomizationSidebarBackgroundStyle,
     CustomizationSidebarListStyle,
-    CustomizationThemeMode,
     type CustomizationThemedColor,
     type CustomizationTint,
     type SiteCustomizationSettings,
@@ -61,7 +61,7 @@ export async function CustomizationRootLayout(props: {
     htmlClassName?: string;
     /** The class name to apply to the body element. */
     bodyClassName?: string;
-    forcedTheme?: CustomizationThemeMode | null;
+    forcedTheme?: CustomizationDefaultThemeMode | null;
     context: GitBookAnyContext;
     children: React.ReactNode;
 }) {
@@ -108,7 +108,7 @@ export async function CustomizationRootLayout(props: {
                     : null,
 
                 // Set the dark/light class statically to avoid flashing and make it work when JS is disabled
-                (forcedTheme ?? customization.themes.default) === CustomizationThemeMode.Dark
+                (forcedTheme ?? customization.themes.default) === CustomizationDefaultThemeMode.Dark
                     ? 'dark'
                     : '',
                 htmlClassName